// React/app/GerAdv_TS/OperadorGruposAgendaOperadores/Interfaces/interface.OperadorGruposAgendaOperadores.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IOperadorGruposAgendaOperadores {
// 202501251
    id: number;
 
	operadorgruposagenda: number,
	operador: number,
}

export interface IOperadorGruposAgendaOperadoresFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IOperadorGruposAgendaOperadoresIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}